requirements.txt
watchdog
tk
pyinstaller
pillow